split bindings
